+++ app/src/main/java/eu/kanade/presentation/more/settings/screen/SettingsAppearanceScreen.kt
@@ -190,6 +190,16 @@ object SettingsAppearanceScreen : SearchableSettings {
                         formattedNow,
                     ),
                 ),
+                Preference.PreferenceItem.SwitchPreference(
+                    preference = uiPreferences.showEpisodeTimestamps(),
+                    title = stringResource(TLMR.strings.pref_show_episode_timestamp),
+                    subtitle = stringResource(TLMR.strings.pref_show_episode_timestamp_summary),
+                ),
+                Preference.PreferenceItem.SwitchPreference(
+                    preference = uiPreferences.showChapterTimestamps(),
+                    title = stringResource(TLMR.strings.pref_show_chapter_timestamp),
+                    subtitle = stringResource(TLMR.strings.pref_show_chapter_timestamp_summary),
+                ),
             ),
         )
     }